refactor routing and user session handling; each user has its own notes page now; update navigation paths and add a different home page for signed out users

+++ src/components/deleteFunction.jsx
@@ -1,9 +1,21 @@
 'use server'
+import { auth } from '@/lib/auth'
+import { headers } from 'next/headers';
 import { PrismaClient } from '@prisma/client';
 
 export async function deleteFunction(id) {
   const prisma = new PrismaClient();
+  const session = await auth.api.getSession({
+    headers: await headers(),
+  })
+  if (!session) {
+    redirect('/sign-in');
+  }
+
   await prisma.note.delete({
-    where: { id: Number(id) }
+    where: {
+      id: Number(id),
+      userId: session.user.id
+    }
     });
 }
